Flesh out basic API for Gc<T>
This commit is fairly large unfortunately. However, much of it will be
familiar as it's been copied from gcmalloc. The key difference is how
we deal with finalization. Unlike gcmalloc, which used `Drop` for both
destruction and finalization, rboehm introduces a new `Finalize`
trait.

Gcmalloc was able to reuse `Drop` because the `drop` method for `GcBox`
would guard the actual `drop` routine based on the object's marking
state. In other words, we would make sure `drop` code wouldn't actually
be called if the object was still alive. This isn't possible with the
Boehm GC as the marking state of an individual object during
finalization is not available. The `Finalize` trait therefore prevents
rustc from recursively calling drop on potential `GcBox`s
indiscriminately. Removing the `Drop` trait from `GcBox` guarantees that
a `GcBox` can now only be finalized explicitly from the finalization
queue by calling `Finalize::finalize()`.

This doesn't affect flexibility from the users perspective. If a user
wants custom drop logic for a `Gc<T>` they would implement `Drop` on `T`
in the same way that they would in gcmalloc.