Require consumer of library to specify the allocator
It's a bit of a shame that this is necessary, since it requires all
consumers of the library to add the following boilerplate:

    #[global_allocator]
    static ALLOCATOR: BoehmAllocator = BoehmAllocator;

However, replacing the global allocator in the library leads to some
rather surprising consequences which are difficult to debug. This is
because the mere inclusion of `rboehm` in the `[dependencies]` section
of a project's `Cargo.toml` is enough to replace the allocator for any
Rust target which has access to it.

This caused a bug in lang_tester because rboehm's allocator was being
used as the global allocator for the harness. This is because testing
and benchmarking harnesses use all dependencies in cargo's
`dependencies` section (as well as `dev-dependencies`). rboehm is not
currently thread-safe, which caused UB when used with lang_tester as it
uses a threadpool to run tests by default.

This fix also triggered a secondary realisation, which is that all
consumers of this library must run unit tests serially until the GC is
made thread-safe.