Fix check and remove Option.
I noticed that this check:

  if layout.size() < tl.size() && layout.align() >= tl.align()

appears to be wrong: it should, I believe, involve "||". While thinking about
that, I came to the conclusion that since new_from_layout is likely to be in
performance critical paths we probably don't want the `Option` involved,
especially since people are only ever unlikely to `unwrap` it. So I've turned
the failure of the check into a `panic`.