Add support for integration with rustc_boehm
This commit introduces a `rustc_boehm` feature. When enabled, rboehm
assumes it is being built with the rustc_boehm fork of rustc and will do
two things: 1) link statically to the BDWGC built with rustc_boehm; and
2) enable optimizations which rely on compiler support. This means that
users who wish to use the rustc_boehm fork for GC will also need to use
the rboehm crate with their project.

Right now, rustc_boehm and rboehm are developed alongside eachother and
roughly 90% of both implementations overlap. This duplication is a pain
to maintain and slows down development.

At the moment, I can't think of a compelling reason to keep a duplicate
of `src/gc.rs` -- the bulk of the Gc API -- inside the standard library
of the forked compiler. There are several advantages to avoiding this:
*much* faster compile times; easier debuggability; easier comparisons;
less duplication, etc.