Allow null fn pointer in `register_finalizer` API
The Boehm GC API allows a null fn pointer to be passed to
`gc_register_finalizer` in order to un-register any existing finalizers.
However, in Rust, it's UB to create a null function pointer. The
solution to this is to wrap the fn pointer type in an `Option`, and pass
a `None` in the null ptr case. At the C ABI level, the `None` is passed
as a null pointer as expected.