Add basic GC stats collection
This provides a way of displaying some useful metrics which will help
debug and profile the GC. Right now, we do not record metrics after each
collection. This is because libgc stops the world uncooperatively. In
other words, there is no "collect" prologue or epilogue to hook into.
These metrics are gated behind the "gc_stats" feature. Performance is
unaffected unless this feature is enabled.