Move boehm-specific stuff into its own crate
To experiment with different collector backends we need to decouple the
Boehm GC code from the more general GC API. I think the most sensible
way to do this is to house the glue code for each backend in its own
nested crate. The GC can then be built with the desired backend by using
the appropriate feature flag. For example, to build with boehm
(currently the only supported option):

    cargo build --features use_boehm

Although this commit has a large diff, it mainly consists of moving
boehm specific code into the boehm/ crate, which is specified as a
dependency in the root Cargo.toml.