Switch to non-ordered finalization
Previously, finalization would happen in topological order.  In other
words, if A points to B, and both are finalizable, A is guaranteed to be
finalized first. This is problematic when cycles between finalizable
objects occur, since the collector can't know which to finalize first.
This means that until now, when cycles occur, finalizers would not be
run. In the context of a Rust GC, this has the potential for memory
leaks.

Switching to non-ordered finalization fixes this problem: there are no
longer any guarantees about which finalizer will run first, which means
cyclic objects can be finalized. However, the danger with non-ordered
finalization is that there must be no references to other GC'd objects
inside the finalizer. If such references exist, then the GC may crash
due to a dangling pointer. I don't believe this is a problem for us, as
we've been careful to avoid this pattern anyway.