Add support for passing object layout info to the GC
At the moment, our GC is fully conservative and scans all allocated
blocks. By providing layout hints, we can speed up marking by telling
the GC that it's safe to ignore certain locations in memory.

While the ultimate goal would be to provide precise layouts for each
allocated object, this is complex. So to start with, an easy win is to
tell the collector that certain objects have a cut-off boundary, after
which there are no more pointers to follow. We call these objects
partially traceable.

To tell the GC that a type `T` is partially traceable, the user must
implement the trait `GcLayout` for `T`. This trait has a method

`layout_info`, where the user can choose one of two options:
    LayoutInfo::Conservative - The default, exact same result as not
    specifying a layout hint at all.

    LayoutInfo::PartiallyTraceable(usize) - Specifies that the collector
    will trace the object up until the provided word boundary. All
    memory after this word is ignored by the collector during marking.

This is a rough first draft. Eventually I see this trait implementation
being sugared behind a procedural macro, but it seems to be a good first
approximation, and also feels naturally expandable to fit precise
layouts in the future too.

Limitations:

    * Currently, the GC only looks at the GcLayout for types that are
    immediately inside a Gc. E.g. When T implements GcLayout, and is
    passed as a type constructor directly to Gc<T>, then the user
    provided layout information is used.

    * PartiallyTraceable can only specify a lower boundary, not a
    "window" inside an object where the pointers may live.