Add #[derive(GcLayout)] proc macro
Implementing layout_info() manually is error-prone (and has caused me
more time than I'd care to admit chasing GC bugs). I think in the vast
number of cases, a user will want to either: a) specify that some fields
in a struct shouldn't be traced, or b) provide a 'trace_end', after
which, fields are ignored by the GC during marking.

This commit sketches out what I think this should look like. It's
explained best with a couple of examples:

    struct S {
        x: usize,
        #[no_trace]
        y: usize,
        z: usize
    }

In this struct, the #[no_trace] attribute tells the GC not to treat `y`
as a pointer, and thus its contents will be ignored by the GC.

    struct T {
        #[trace_end]
        x: usize,
        y: usize,
        z: usize
    }

In this struct, the #[trace_end] attribute tells the GC that `x` is the
last field to be traced, and that it should ignore all fields after it.
In other words, only `T.x` is traced by the GC.

The procedural macro generates a function body for
GcLayout::layout_info() which constructs a bitmap representing the
struct's layout. For most structs, where field sizes are known at
compile-time, the body of layout_info() will be constant-folded into a
single return instruction.