Reimplement fmt::Debug and fmt::Pointer for Gc<T>
The default implementation of fmt::Debug showed only the pointer
of the Gc'd value, and does not allow the inner value to be debug
printed by itself, without some sort of newtype shim. An
implementation was written which instead debug formats the inner
value, in the same way as other box values such as Box<T> or Rc<T>.

The fmt::Pointer trait was also implemented so the Gc value can
be formatted as a pointer if desired, so no functionality is lost.