Remove requirement that `T: Sync`
This was included originally because of concerns about finalizing a
non-Sync `T` value off-thread.

If a value is `Sync`, it means that it is safe for two threads to access
its data simultaneuosly: in other words, it has some synchronization
guarantees. A common example is `RefCell`, which is safe to send between
threads, but is marked `!Sync` because the dynamic borrow checking is
not atomic. Finalizing `!Sync` values no longer problematic, because
even though finalizers are run off-thread, they will run when the object
is dead [1],  so there is no chance of a data race.

[1]: https://github.com/softdevteam/rustgc/pull/30